Cablea Pedido con UnidadDeVenta y Festival con UnidadDeVenta
Pedido ya no guarda festival directo: se deriva via unidad.festival
(decision confirmada con la catedra, UnidadDeVenta ya sabe a que
festival pertenece). Se suma el campo festival a UnidadDeVenta
(nullable por ahora) para que esa derivacion sea posible.

De paso se corrigen dos bugs en UnidadDeVentaDao/ABM que aparecieron
al ejercitar esta relacion: traer(long) no coincidia con el tipo real
de la PK (int), y las consultas no traian festival con join fetch, lo
que rompia con LazyInitializationException apenas se intentaba
imprimir una unidad fuera de la sesion que la trajo.

+++ src/negocio/UnidadDeVentaABM.java
@@ -36,7 +36,7 @@ public class UnidadDeVentaABM {
 		}
 	}
 
-	public UnidadDeVenta traer(long id) {
+	public UnidadDeVenta traer(int id) {
 		return dao.traer(id);
 	}
 
@@ -56,7 +56,7 @@ public class UnidadDeVentaABM {
 		dao.actualizar(u);
 	}
 
-	public void eliminar(long id) throws Exception {
+	public void eliminar(int id) throws Exception {
 		UnidadDeVenta u = dao.traer(id);
 		if (u == null) {
 			throw new Exception("No existe la unidad de venta con id " + id);
